use the disabled palette to render husks
closes #4631
