+++ CHANGELOG
@@ -12,6 +12,7 @@ NEW:
 			Added skirmish mode to RA and D2k to complement TD's skirmish mode.
 			Added an Extras submenu for miscellaneous game extras.
 		Engineers can now regain control over husks.
+		Husks are now rendered with a black overlay.
 		A player's units, and allied units, now move out of the way when blocking production facilities.
 		Added cheat button to grow map resources.
 		Fixed units staying selected and contributing to control groups when becoming cloaked or hidden in fog.
@@ -105,6 +106,7 @@ NEW:
 		Mods that use custom TileSize must specify both width and height.
 		If you spot black tiles in your Dune 2000 ARRAKIS maps, replace them with the remaining sand and rock tiles. Go to Map → Fix Open Areas to randomize them.
 		The TestFile check in mod.yaml has been renamed to TestFiles (plural!) and now supports a comma-separated list of assets that are required to load the game.
+		DisabledOverlay has been split from RenderBuilding. Use it together with RequiresPower and CanPowerDown for buildings or directly with Husk.
 	Packaging:
 		Removed portable install option from Windows installer as the game left without write access breaks content download and error log generation.
 		Added HTML documentation to the Windows installer.
