+++ OpenRA.Mods.RA/Burns.cs
@@ -32,6 +32,7 @@ namespace OpenRA.Mods.RA
 			Info = info;
 
 			var anim = new Animation("fire", () => 0);
+			anim.IsDecoration = true;
 			anim.PlayRepeating(Info.Anim);
 			self.Trait<RenderSprites>().anims.Add("fire", anim);
 		}
