+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -495,6 +495,7 @@
     <Compile Include="Buildings\Demolishable.cs" />
     <Compile Include="Widgets\Logic\DisconnectWatcherLogic.cs" />
     <Compile Include="Activities\FlyFollow.cs" />
+    <Compile Include="Modifiers\DisabledOverlay.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\OpenRA.FileFormats\OpenRA.FileFormats.csproj">
