+++ mods/cnc/rules/defaults.yaml
@@ -523,6 +523,7 @@
 	BelowUnits:
 	BodyOrientation:
 	LuaScriptEvents:
+	DisabledOverlay:
 
 ^HelicopterHusk:
 	Inherits: ^Husk
