+++ mods/d2k/rules/atreides.yaml
@@ -61,6 +61,7 @@ PALACEA:
 		SelectTargetSound:
 		FlareType:
 	CanPowerDown:
+	DisabledOverlay:
 	RequiresPower:
 	SupportPowerChargeBar:
