+++ mods/ra/rules/defaults.yaml
@@ -438,6 +438,7 @@
 		CaptureThreshold: 1.0
 	TransformOnCapture:
 		ForceHealthPercentage: 25
+	DisabledOverlay:
 
 ^HelicopterHusk:
 	Inherits: ^Husk
